+++ api/tests/Pest.php
@@ -1,5 +1,6 @@
 <?php
 
+use App\Models\User;
 use Illuminate\Foundation\Testing\RefreshDatabase;
 use Tests\TestCase;
 
@@ -17,3 +18,11 @@ use Tests\TestCase;
 pest()->extend(TestCase::class)
     ->use(RefreshDatabase::class)
     ->in('Feature');
+
+function actingAsUser(): User
+{
+    $user = User::factory()->create();
+    test()->actingAs($user, 'sanctum');
+
+    return $user;
+}